Update to remove SmallSharp and add some folder structure. Example 07 still a WIP
