+++ src/Vts.Scripting/Program.cs
@@ -1,0 +1,30 @@
+﻿using Vts.Scripting.MonteCarlo;
+
+/// <summary>
+/// Top program used to execute the desired demo script(s)
+/// </summary>
+public class Program
+{
+    /// <summary>
+    /// Main function that executes the desired demo script(s)
+    /// </summary>
+    /// <param name="args"></param>
+    public static void Main(string[] args)
+    {
+        //RunAllMonteCarloDemos();
+
+        MC07_pMCInversion.RunDemo();
+    }
+
+    public static void RunAllMonteCarloDemos()
+    {
+        MC01_ROfRhoSimple.RunDemo();
+        MC02_DAWvsCAW.RunDemo();
+        MC03_ROfRhoFullCustomization.RunDemo();
+        MC04_N1000vsN100.RunDemo();
+        MC05_PostProcessor.RunDemo();
+        MC06_pMCPostProcessor.RunDemo();
+        MC07_pMCInversion.RunDemo();
+        MC10_ROfFx.RunDemo();
+    }
+}
