--- src/Vts.Scripting/Properties/launchSettings.json
@@ -1,25 +1,0 @@
-{
-  "profiles": {
-    "MonteCarlo-01-ROfRhoSimple.cs": {
-      "commandName": "Project"
-    },
-    "MonteCarlo-02-DAWvsCAW.cs": {
-      "commandName": "Project"
-    },
-    "MonteCarlo-03-ROfRhoFullCustomization.cs": {
-      "commandName": "Project"
-    },
-    "MonteCarlo-04-N1000vsN100.cs": {
-      "commandName": "Project"
-    },
-    "MonteCarlo-05-PostProcessor.cs": {
-      "commandName": "Project"
-    },
-    "MonteCarlo-06-pMCPostProcessor.cs": {
-      "commandName": "Project"
-    },
-    "MonteCarlo-10-ROfFx.cs": {
-      "commandName": "Project"
-    }
-  }
-}
